Add some tests before the end of the day
